test: dedupe code using macro
